USP now checked against R, speed up some tests

+++ src/resample/permutation.py
@@ -145,19 +145,27 @@ def usp(
 
     rng = _normalize_rng(random_state)
 
-    w = np.asarray(w, dtype=int)
+    w = np.asarray(w, dtype=float)
     if w.ndim != 2:
         raise ValueError("w must be two-dimensional")
     wx = np.sum(w, axis=1)
     wy = np.sum(w, axis=0)
-    n = np.sum(wx)
+    n = int(np.sum(wx))
 
-    m = np.outer(wx, wy).astype(float) / n
+    m = np.outer(wx, wy) / n
 
     f1 = 1.0 / (n * (n - 3))
     f2 = 4.0 / (n * (n - 2) * (n - 3))
 
     t = _usp(f1, f2, w, m)
+
+    # TODO: The shuffling algorithm used here has O(N) complexity in space and time
+    # where N is the total number of entries in the input array. The R implementation
+    # uses Patefield's algorithm, see https://rdrr.io/r/stats/r2dtable.html, which has
+    # O(K) complexity in space and time, where K is the total number of cells in the
+    # table. For N >> K, which can easily happen in high-energy physics, the latter
+    # will be dramatically faster. There seems to be no Python implementation of
+    # Patefield's algorithm right now.
 
     # generate x,y index arrays
     xmap = np.empty(n, dtype=int)
@@ -171,8 +179,8 @@ def usp(
             k += wij
 
     # For Type I error probabilities to hold theoretically, the number of bootstrap
-    # samples drawn may not the depend on the data (comment by Richard Samworth).
-    # So we compute the required number of samples with the worst-case p=0.5.
+    # samples may not the depend on the data (comment by Richard Samworth), so not on
+    # the actual pvalue for this data set. So we compute it with the worst-case p=0.5.
     n = min(max_size, int(0.25 / precision**2)) if precision > 0 else max_size
     ts = np.empty(n)
     for b in range(n):
